pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["hungrytime <nothing>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
strawberry-graphql = {extras = ["debug-server"], version = "^0.237.3"}
strawberry-graphql-django = "^0.47.1"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
